[FE Modularization] Move `FieldPanel` to visualizations
