+++ frontend/src/metabase/visualizations/components/settings/ChartSettingTableColumns/FieldPanel/types.ts
