+++ frontend/src/metabase/visualizations/components/settings/ChartSettingTableColumns/FieldPanel/utils.ts
